requirements.txt
urllib3
selenium~=3.141.0
natsort~=8.2.0
pathlib~=1.0.1
python-dotenv~=0.21.0
googlemaps~=4.7.3
bs4~=0.0.1
beautifulsoup4~=4.9.3
xlrd~=2.0.1
requests~=2.28.1